+++ tests/integration/example/example_test.go
@@ -44,9 +44,6 @@ func Example() {
 	// replace the logger by testing values in a real test case (e.g. log.NewTestLogger(t))
 	logger := log.NewNopLogger()
 
-	cms := integration.CreateMultiStore(keys, logger)
-	newCtx := sdk.NewContext(cms, true, logger)
-
 	// gomock initializations
 	ctrl := gomock.NewController(&testing.T{})
 	acctsModKeeper := authtestutil.NewMockAccountsModKeeper(ctrl)
@@ -78,7 +75,6 @@ func Example() {
 
 	// create the application and register all the modules from the previous step
 	integrationApp := integration.NewIntegrationApp(
-		newCtx,
 		logger,
 		keys,
 		encodingCfg.Codec,
@@ -148,9 +144,6 @@ func Example_oneModule() {
 	// replace the logger by testing values in a real test case (e.g. log.NewTestLogger(t))
 	logger := log.NewLogger(io.Discard)
 
-	cms := integration.CreateMultiStore(keys, logger)
-	newCtx := sdk.NewContext(cms, true, logger)
-
 	// gomock initializations
 	ctrl := gomock.NewController(&testing.T{})
 	acctsModKeeper := authtestutil.NewMockAccountsModKeeper(ctrl)
@@ -177,7 +170,6 @@ func Example_oneModule() {
 
 	// create the application and register all the modules from the previous step
 	integrationApp := integration.NewIntegrationApp(
-		newCtx,
 		logger,
 		keys,
 		encodingCfg.Codec,