+++ tests/integration/gov/keeper/keeper_test.go
@@ -4,8 +4,12 @@ import (
 	"context"
 	"testing"
 
+<<<<<<< HEAD
 	"github.com/golang/mock/gomock"
 	"gotest.tools/v3/assert"
+=======
+	"go.uber.org/mock/gomock"
+>>>>>>> 2d04a1af2 (fix(testutil/integration): use only one context in integration test framework (#22616))
 
 	"cosmossdk.io/core/appmodule"
 	"cosmossdk.io/log"
@@ -63,10 +67,6 @@ func initFixture(tb testing.TB) *fixture {
 	cdc := encodingCfg.Codec
 
 	logger := log.NewTestLogger(tb)
-	cms := integration.CreateMultiStore(keys, logger)
-
-	newCtx := sdk.NewContext(cms, true, logger)
-
 	authority := authtypes.NewModuleAddress(types.ModuleName)
 
 	maccPerms := map[string][]string{
@@ -111,8 +111,6 @@ func initFixture(tb testing.TB) *fixture {
 		authority.String(),
 	)
 
-	assert.NilError(tb, bankKeeper.SetParams(newCtx, banktypes.DefaultParams()))
-
 	router := baseapp.NewMsgServiceRouter()
 	queryRouter := baseapp.NewGRPCQueryRouter()
 	consensusParamsKeeper := consensusparamkeeper.NewKeeper(cdc, runtime.NewEnvironment(runtime.NewKVStoreService(keys[consensusparamtypes.StoreKey]), log.NewNopLogger(), runtime.EnvWithQueryRouterService(queryRouter), runtime.EnvWithMsgRouterService(router)), authtypes.NewModuleAddress("gov").String())
@@ -120,10 +118,6 @@ func initFixture(tb testing.TB) *fixture {
 	stakingKeeper := stakingkeeper.NewKeeper(cdc, runtime.NewEnvironment(runtime.NewKVStoreService(keys[stakingtypes.StoreKey]), log.NewNopLogger()), accountKeeper, bankKeeper, consensusParamsKeeper, authority.String(), addresscodec.NewBech32Codec(sdk.Bech32PrefixValAddr), addresscodec.NewBech32Codec(sdk.Bech32PrefixConsAddr), runtime.NewContextAwareCometInfoService())
 
 	poolKeeper := poolkeeper.NewKeeper(cdc, runtime.NewEnvironment(runtime.NewKVStoreService(keys[pooltypes.StoreKey]), log.NewNopLogger()), accountKeeper, bankKeeper, authority.String())
-
-	// set default staking params
-	err := stakingKeeper.Params.Set(newCtx, stakingtypes.DefaultParams())
-	assert.NilError(tb, err)
 
 	// Create MsgServiceRouter, but don't populate it before creating the gov
 	// keeper.
@@ -140,12 +134,9 @@ func initFixture(tb testing.TB) *fixture {
 		keeper.DefaultConfig(),
 		authority.String(),
 	)
-	assert.NilError(tb, govKeeper.ProposalID.Set(newCtx, 1))
 	govRouter := v1beta1.NewRouter()
 	govRouter.AddRoute(types.RouterKey, v1beta1.ProposalHandler)
 	govKeeper.SetLegacyRouter(govRouter)
-	err = govKeeper.Params.Set(newCtx, v1.DefaultParams())
-	assert.NilError(tb, err)
 
 	authModule := auth.NewAppModule(cdc, accountKeeper, acctsModKeeper, authsims.RandomGenesisAccounts, nil)
 	bankModule := bank.NewAppModule(cdc, bankKeeper, accountKeeper)
@@ -153,7 +144,7 @@ func initFixture(tb testing.TB) *fixture {
 	govModule := gov.NewAppModule(cdc, govKeeper, accountKeeper, bankKeeper, poolKeeper)
 	consensusModule := consensus.NewAppModule(cdc, consensusParamsKeeper)
 
-	integrationApp := integration.NewIntegrationApp(newCtx, logger, keys, cdc,
+	integrationApp := integration.NewIntegrationApp(logger, keys, cdc,
 		encodingCfg.InterfaceRegistry.SigningContext().AddressCodec(),
 		encodingCfg.InterfaceRegistry.SigningContext().ValidatorAddressCodec(),
 		map[string]appmodule.AppModule{